[refactor] Optional 적용
