+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,12 +1,12 @@
-package gift.dao;
+package gift.repository;
 
 import java.util.List;
-
+import java.util.Optional;
 import gift.domain.Product;
 
-public interface ProductDao {
+public interface ProductRepository {
 	void save(Product product);
-	Product findById(Long id);
+	Optional<Product> findById(Long id);
 	List<Product> findAll();
 	void update(Product product);
 	void deleteById(Long id);
